requirements.txt
slack-bolt>=1.15,<2
notion-client>=1,<2
python-dotenv
